chore: refactored operation review
