+++ bciers/apps/reporting/src/data/jsonSchema/reviewOperationInformationText.tsx
@@ -23,3 +23,6 @@ export const purposeNote = (
     </Box>
   </Paper>
 );
+
+export const reportTypeHelperText =
+  "Simple Reports are submitted by reporting operations that previously emitted greater than or equal to 10 000 tCO2e of attributable emissions in a reporting period, but now emit under 10 000 tCO2e of attributable emissions and have an obligation to continue reporting emissions for three consecutive reporting periods. This report type is not applicable for any operations that received third party verification in the immediately preceding reporting period, and is not applicable for opted-in operations.";
